Move some components around
Why:

There were several page-level components inside the src/components directory, and they shouldn't be there, especially for purposes of the style guide.

This change addresses the need by:

Creating a new `src/pages` directory and putting the pages in there. This will be easier to maintain in the long run.